# PERF: lib.generate_slices