@jreback 2021-06-18
@meeseeksdev backport 1.3.x